Dynamic jumping and turn speeds
Jumping is more dynamic now, it uses both jumping animations instead of one. It checks if the player direction variable is not zero to determine the jump type. And the turn speed is automatically checked if the player is on the floor or not, you rotate faster on solid ground than if you were in the air.